New loop color propagation pass (#277)
This adds a new coloring propagation policy for while loops when we use a custom schedule that first goes through colors from shallowest to deepest and propagates colors forward to users, rewriting when necessary, then goes from deepest to shallowest and propagates to operands. This should help with keeping user tasks together.

Signed-off-by: Murali Andoorveedu <mandoorveedu@nvidia.com>